backend/services/groq.py
import os
from dotenv import load_dotenv

load_dotenv()

GROQ_API_KEY = os.getenv("GROQ_API_KEY")

def accelerate_inference(model_inputs):
    # Placeholder per utilizzare Groq se disponibile
    # Possibile utilizzo: velocizzare l'inferenza LLM o immagine
    return model_inputs
