backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.0
openai==1.57.0
packaging==24.2
pip-tools==7.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pyproject_hooks==1.2.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wheel==0.45.1
